Per-turbine Collision Risk Model for predicting avian collision with on-shore and off-shore wind turbines

This package contains a stochastic avian - wind turbine collision risk model that enables distance sampling and per-turbine calculations.

Package Content

Index: This package was not yet installed at build time.

Author

Elizabeth Stark [aut, cre] (ORCID: <https://orcid.org/0009-0002-6337-5021>), Muir Stuart [aut], Yandell Peggy [aut], Akila Sahinya [ctb], Shao Ray [ctb], Biosis [cph] (Package uses and adapts turbine geometry methods sourced with permission from Biosis' Collision Risk Model), Symbolix [cph]

Maintainer

Elizabeth Stark <estark@symbolix.com.au>